JWT Token

+++ lib/repositories/preferences_repository.dart
@@ -17,7 +17,7 @@ class PreferencesRepository {
     if (json == null) {
       return [];
     } else {
-      return jsonDecode(json);
+      return jsonDecode(json).cast<int>();
     }
   }
